# source/backend/api/openapi.yaml
openapi: 3.0.0
info:
  title: UniReview API
  version: 1.0.0
servers:
  - url: /api
    description: Base URL for API endpoints
components:
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
        moderatore:
          type: boolean
        nomeUtente:
          type: string
        bannedUntil:
          type: string
          format: date
        linguaUI:
          type: boolean
        temaUI:
          type: boolean

    Review:
      type: object
      properties:
        professore:
          type: string
        esame:
          type: string
        valutazioneProfessore:
          type: number
          minimum: 1
          maximum: 5
        valutazioneFattibilita:
          type: number
          minimum: 1
          maximum: 5
        valutazioneMateriale:
          type: number
          minimum: 1
          maximum: 5
        testo:
          type: string
        tentativo:
          type: number
        voto:
          type: number
        frequenza:
          type: string
          enum: ["0%", "<50%", ">50%"]
        anonima:
          type: boolean

    Exam:
      type: object
      properties:
        professore:
          type: string
        esame:
          type: string

paths:
  /user:
    get:
      summary: Restituisce tutti gli utenti (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      responses:
        "200":
          description: Successo
        "500":
          description: Errore del server

  /user/{email}:
    get:
      summary: Restituisce un utente per email
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Utente non trovato
        "500":
          description: Errore del server

    patch:
      summary: Aggiorna un utente per email (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Utente aggiornato con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Utente non trovato
        "500":
          description: Errore del server

    delete:
      summary: Elimina un utente per email (solo moderatore o utente stesso)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Utente eliminato con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Utente non trovato
        "500":
          description: Errore del server

  /user/theme:
    put:
      summary: Modifica Tema UI utente attuale
      tags:
        - Protetto
      security:
        - JWT: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                themeUi:
                  type: boolean
                  description: Nuovo valore per il tema UI
      responses:
        "200":
          description: Tema UI modificato con successo
        "500":
          description: Errore del server

  /user/language:
    put:
      summary: Modifica Lingua UI utente attuale
      tags:
        - Protetto
      security:
        - JWT: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                languageUi:
                  type: boolean
                  description: Nuovo valore per la lingua UI
      responses:
        "200":
          description: Lingua UI modificata con successo
        "500":
          description: Errore del server

  /user/{email}/revoke-moderator:
    patch:
      summary: Torna utente standard da moderatore (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: email dell'utente da modificare
      responses:
        "200":
          description: Utente modificato con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Utente non trovato
        "500":
          description: Errore del server

  /user/{email}/promote-moderator:
    patch:
      summary: Promuovi utente standard a moderatore (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: email dell'utente da modificare
      responses:
        "200":
          description: Utente modificato con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Utente non trovato
        "500":
          description: Errore del server

  /user/{email}/ban:
    patch:
      summary: Banna un utente per un numero specifico di giorni (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: email dell'utente da bannare
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  minimum: 1
                  description: Numero di giorni per il ban
      responses:
        "200":
          description: Utente modificato con successo
        "400":
          description: Input non valido
        "403":
          description: Accesso non autorizzato
        "404":
          description: Utente non trovato
        "500":
          description: Errore del server

  /user/{email}/unban:
    patch:
      summary: Rimuove il ban di un utente (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: email dell'utente da cui rimuovere il ban
      responses:
        "200":
          description: Utente modificato con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Utente non trovato
        "500":
          description: Errore del server

 
  /review:
    get:
      summary: Restituisce tutte le recensioni
      responses:
        "200":
          description: Successo
        "500":
          description: Errore del server
  
    post:
      summary: Componi nuova recensione (richiede autenticazione)
      tags:
        - Protetto
      security:
        - JWT: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review"
      responses:
        "201":
          description: Recensione creata con successo
        "400":
          description: Errore nell'inserimento dei dati"
        "500":
          description: Errore del server
  
  /review/{reviewId}:
    patch:
      summary: Modifica recensione esistente per ID (richiede autenticazione)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review"
      responses:
        "200":
          description: Recensione modificata con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Recensione non trovata
        "500":
          description: Errore del server

    delete:
      summary: Rimuovi recensione per ID (richiede autenticazione)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Recensione rimossa con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Recensione non trovata
        "500":
          description: Errore del server


  /review/{email}:
    get:
      summary: Ottieni tutte le recensioni dell'utente specificato
      responses:
        "200":
          description: Recensioni ottenute con successo
        "404":
          description: Nessuna recensione trovata
        "500":
          description: Errore del server

  /review/professore/{professore}:
    get:
      summary: Ottieni tutte le recensioni del professore specificato
      responses:
        "200":
          description: Recensioni ottenute con successo
        "404":
          description: Nessuna recensione trovata
        "500":
          description: Errore del server

  /review/esame/{esame}:
    get:
      summary: Ottieni tutte le recensioni dell'esame specificato
      responses:
        "200":
          description: Recensioni ottenute con successo
        "404":
          description: Nessuna recensione trovata
        "500":
          description: Errore del server

  /exam:
    get:
      summary: Restituisce tutti gli esami
      responses:
        "200":
          description: Successo
        "500":
          description: Errore del server"
    
    post:
      summary: Crea un nuovo esame (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Exam"
      responses:
        "201":
          description: Utente modificato con successo
        "400":
          description: Errore nell'inserimento dei dati
        "403":
          description: Accesso non autorizzato
        "500":
          description: Errore del server

    delete:
      summary: Cancella un esame (solo moderatore)
      tags:
        - Protetto
      security:
        - JWT: []
      parameters:
        - name: esame
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Esame cancellato con successo
        "403":
          description: Accesso non autorizzato
        "404":
          description: Esame non trovato
        "500":
          description: Errore del server
      
  /exam/esame/{esame}:
      get:
        summary: Restituisce una lista di esami per nome
        parameters:
          - name: esame
            in: path
            required: true
            schema:
              type: string
        responses:
          "200":
            description: Successo
          "404":
            description: Nessun esame trovato
          "500":  
            description: Errore del server"

  /exam/professore/{professore}:
      get:
        summary: Restituisce tutti gli esami per professore
        parameters:
          - name: professore
            in: path
            required: true
            schema:
              type: string
        responses:
          "200":
            description: Successo
          "404":
            description: Nessun esame trovato
          "500":  
            description: Errore del server"

  /send-email:
    post:
      summary: Invia una segnalazione per email (richiede autenticazione)
      tags:
        - Protetto
      security:
        - JWT: []
      requestBody:
        required: true
        content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                    example: "Ho bisogno di aiuto con la mia recensione."
                    description: Il testo dell'email da inviare.
                required:
                  - text
      responses:
        '200':
          description: Email inviata con successo
        '400':
          description: Richiesta non valida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Il testo dell'email è obbligatorio"
        '401':
          description: Non autorizzato
        '500':
          description: Errore del server